Preparations for 0.0.12-Windows-Alpha
Swapping to some more advanced and secure features. Here is what I've added:

- Console automatically goes to full screen when booted up.
- Console will clear the moment the application is booted up.
- No errors when compiling! :smiley:
- Commands are no longer tied to a number but a string instead.
- Added clear command and printStartMenu command.
- Updated the help command.

One feature that I'm totally going to add either after school or when I get home is turning the string input into a hashed number. This will make it sooo much better to read in the code and will also allow us to remove that really gross looking if-else chain that checks for the different commands.

That's all. :trollface:

Coding in Class,
- Jack Spencer