0.0.12-Windows-Alpha preparations real?!?
Moving a lot of files to different areas, and I've also added a cool little feature where you can get a hash.

Too bad its a bit broken. :/

File structure has been updated, so that core.c will stay in the root of the windows folder, and the other files will be put into the different folders, such as /lib and /bin.

I'm also renaming all of the functions prefixes to the specific file they are in. For example: a function called clearScreen in the file util.c would be titled UTIL_clearScreen. Looks better anyways.

This is already a big step in the direction of future proof code. Once I get the program working on this new infrastructure I'll start working on some more features. If you look in the makima.c file and the util.c file you can see some of the stuff I'm planning on adding.

This project is slowly but surely getting better, and I'm really glad that it's looking this good at this point so far.

Currently near a chess board,
- Jack Spencer